feat: disable autodeploy by default
this change is at the request of product. publishing the starter site creates
streams for DM entities before they exist. when the DM runs it causes a spike
as each entity is created. we should be able to revert at some point when product
has finished re-architecting the DM

J/T=none

TEST=none